# area_back_end/web-services/docs/swagger.yaml
definitions:
...
  login.Code:
    properties:
      code:
        type: string
    type: object
  login.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  login.LoginCredentials:
    properties:
      password:
        example: password123
        type: string
      username:
        example: user@example.com
        type: string
    type: object
  login.LoginErrorResponse:
    properties:
      message:
        example: Incorrect credentials
        type: string
    type: object
  login.LoginResponse:
    properties:
      message:
        type: string
      token:
        type: string
    type: object
  login.LoginSuccessResponse:
    properties:
      message:
        example: Login successful
        type: string
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
  login.RegisterCredentials:
    properties:
      email:
        example: newuser@example.com
        type: string
      password:
        example: securePassword123
        type: string
      username:
        example: newuser
        type: string
    type: object
  login.RegisterErrorResponse:
    properties:
      message:
        example: Username already taken
        type: string
    type: object
  login.RegisterSuccessResponse:
    properties:
      message:
        example: Account registered successfully
        type: string
    type: object
info:
  contact: {}
paths:
...
  /login:
    post:
      consumes:
      - application/json
      description: Authenticate user and provide a token upon successful authentication.
      parameters:
      - description: API Key
        in: header
        name: x-api-key
        required: true
        type: string
      - description: User Credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/login.LoginCredentials'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            $ref: '#/definitions/login.LoginSuccessResponse'
        "400":
          description: Invalid username or password
          schema:
            $ref: '#/definitions/login.LoginErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: User login
      tags:
      - Auth
  /login/discord:
    post:
      consumes:
      - application/json
      description: Login or register a new user using discord OAuth by providing a
        discord code.
      parameters:
      - description: API key for access control
        in: header
        name: x-api-key
        required: true
        type: string
      - description: Bearer token if the user is already logged in
        in: header
        name: Bearer
        type: string
      - description: discord code for OAuth
        in: body
        name: code
        required: true
        schema:
          $ref: '#/definitions/login.Code'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful with token
          schema:
            $ref: '#/definitions/login.LoginResponse'
        "400":
          description: Bad request if the code is not provided or invalid
          schema:
            $ref: '#/definitions/login.ErrorResponse'
        "401":
          description: Unauthorized if the API key is not valid
          schema:
            $ref: '#/definitions/login.ErrorResponse'
        "500":
          description: Internal Server Error when login fails
          schema:
            $ref: '#/definitions/login.ErrorResponse'
      summary: Login with discord OAuth
      tags:
      - Auth
  /login/github:
    post:
      consumes:
      - application/json
      description: Login or register a new user using GitHub OAuth by providing a
        GitHub code.
      parameters:
      - description: API key for access control
        in: header
        name: x-api-key
        required: true
        type: string
      - description: Bearer token if the user is already logged in
        in: header
        name: Bearer
        type: string
      - description: GitHub code for OAuth
        in: body
        name: code
        required: true
        schema:
          $ref: '#/definitions/login.Code'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful with token
          schema:
            $ref: '#/definitions/login.LoginResponse'
        "400":
          description: Bad request if the code is not provided or invalid
          schema:
            $ref: '#/definitions/login.ErrorResponse'
        "401":
          description: Unauthorized if the API key is not valid
          schema:
            $ref: '#/definitions/login.ErrorResponse'
        "500":
          description: Internal Server Error when login fails
          schema:
            $ref: '#/definitions/login.ErrorResponse'
      summary: Login with GitHub OAuth
      tags:
      - Auth
...
  /register:
    post:
      consumes:
      - application/json
      description: Register a new user with a username, email, and password.
      parameters:
      - description: API Key
        in: header
        name: x-api-key
        required: true
        type: string
      - description: User Registration Data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/login.RegisterCredentials'
      produces:
      - application/json
      responses:
        "201":
          description: Account registered successfully
          schema:
            $ref: '#/definitions/login.RegisterSuccessResponse'
        "400":
          description: Invalid input data
          schema:
            $ref: '#/definitions/login.RegisterErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: User registration
      tags:
      - Auth
swagger: "2.0"